tests/wpt/tests/html/dom/render-blocking/document-render-blocking-partial.tentative.html

<!DOCTYPE html>
  <html blocking=render>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/utils.js"></script>
  <title>`blocking=render` defers frames until the attribute is set</title>
  <script>
    assert_implements(document.documentElement.blocking, "no blocking attribute");

    promise_test(() => {
      return new Promise((resolve, reject) => {
        requestAnimationFrame(() => {
          if (document.getElementById("last"))
            reject();
          else
            resolve();
        });
      });
    }, "blocking defers frames until removed");
  </script>
  </head>
  <body>
    <div id="first"></div>
    <script>
      jankMany(100, 10);
      document.documentElement.blocking="";
    </script>
    <div id="second"></div>
    <script>
      jankMany(100, 10);
    </script>
    <div id="last"></div>
  </body>
</html>
